requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mfrc522-python==0.0.7
RPi.GPIO==0.7.1
spidev==3.6
Werkzeug==3.1.3
